// src/app/store/associate/associate.types.ts
export enum AssociateActionsType {
  LOAD_ASSOCIATE = '[Associate Page]Load Associate',
  LOAD_ASSOCIATE_SUCCESS = '[Associate Page]Load Associate Success',
  LOAD_ASSOCIATE_FAIL = '[Associate Page]Load Associate Fail',
  ADD_ASSOCIATE = '[Associate Page]Add Associate',
  ADD_ASSOCIATE_SUCCESS = '[Associate Page]Add Associate Success',
  UPDATE_ASSOCIATE = '[Associate Page]Update Associate',
  UPDATE_ASSOCIATE_SUCCESS = '[Associate Page]Update Associate Success',
  DELETE_ASSOCIATE = '[Associate Page]Delete Associate',
  DELETE_ASSOCIATE_SUCCESS = '[Associate Page]Delete Associate Success',
  GET_ASSOCIATE = '[Associate Page]Get Associate',
  GET_ASSOCIATE_SUCCESS = '[Associate Page]Get Associate Success',
  OPEN_POPUP = '[Associate Page]Open Popup',
}